+++ application/external/src/test/kotlin/io/github/siyual_park/application/external/annoration/ControllerTest.kt
@@ -1,7 +1,6 @@
-package io.github.siyual_park.application.external.controller
+package io.github.siyual_park.application.external.annoration
 
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient
-import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
 @AutoConfigureWebTestClient
