+++ application/external/src/test/kotlin/io/github/siyual_park/application/external/annoration/SpringBootTest.kt
@@ -1,0 +1,8 @@
+package io.github.siyual_park.application.external.annoration
+
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
+
+@SpringBootTest
+@ActiveProfiles("test")
+annotation class SpringBootTest
