+++ application/external/src/test/resources/application-test.properties
@@ -1,0 +1,1 @@
+logging.level.org.springframework.r2dbc=DEBUG
